+++ include/drivers/keyboard.h
@@ -20,6 +20,9 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #ifndef KEYBOARD_H
 #define KEYBOARD_H
 
+#define KEY_UP    0x11
+#define KEY_DOWN  0x12
+
 #define KBD_LEN        58
 #define KBD_BUFFER_LEN 1024