+++ src/drivers/keyboard.cpp
@@ -20,7 +20,6 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #include <stdio.h>
 
 #include "architecture/io.h"
-#include "drivers/terminal.h"
 
 #include "drivers/keyboard.h"